+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - add .lazy_attr_accessor
 - .lazy_attr_reader and .lazy_attr_accessor return the same value as their normal counterparts
+- #eager_load_attributes! returns the names of only the  attributes which were eager-loaded
 
 ## 0.1.0